pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "tictactoe-attempt"
version = "0.1.0"
[tool.poetry.dependencies]
ansicolors = "^1.1"
python = "^3.7"
[tool.poetry.dev-dependencies]
